Final touch: move service to MixPhpHtmlDecorator
