+++ src/Rector/AbstractRector.php
@@ -28,11 +28,11 @@ use Rector\Core\Exception\ShouldNotHappenException;
 use Rector\Core\FileSystem\FilePathHelper;
 use Rector\Core\Logging\CurrentRectorProvider;
 use Rector\Core\NodeDecorator\CreatedByRuleDecorator;
+use Rector\Core\NodeDecorator\MixPhpHtmlDecorator;
 use Rector\Core\PhpParser\Comparing\NodeComparator;
 use Rector\Core\PhpParser\Node\BetterNodeFinder;
 use Rector\Core\PhpParser\Node\NodeFactory;
 use Rector\Core\PhpParser\Node\Value\ValueResolver;
-use Rector\Core\PhpParser\Printer\MixPhpHtmlPrinter;
 use Rector\Core\ProcessAnalyzer\RectifiedAnalyzer;
 use Rector\Core\Provider\CurrentFileProvider;
 use Rector\Core\ValueObject\Application\File;
@@ -116,7 +116,7 @@ CODE_SAMPLE;
 
     private DocBlockUpdater $docBlockUpdater;
 
-    private MixPhpHtmlPrinter $mixPhpHtmlPrinter;
+    private MixPhpHtmlDecorator $mixPhpHtmlDecorator;
 
     #[Required]
     public function autowire(
@@ -141,7 +141,7 @@ CODE_SAMPLE;
         RectorOutputStyle $rectorOutputStyle,
         FilePathHelper $filePathHelper,
         DocBlockUpdater $docBlockUpdater,
-        MixPhpHtmlPrinter $MixPhpHtmlPrinter
+        MixPhpHtmlDecorator $mixPhpHtmlDecorator
     ): void {
         $this->nodesToRemoveCollector = $nodesToRemoveCollector;
         $this->nodeRemover = $nodeRemover;
@@ -164,7 +164,7 @@ CODE_SAMPLE;
         $this->rectorOutputStyle = $rectorOutputStyle;
         $this->filePathHelper = $filePathHelper;
         $this->docBlockUpdater = $docBlockUpdater;
-        $this->mixPhpHtmlPrinter = $MixPhpHtmlPrinter;
+        $this->mixPhpHtmlDecorator = $mixPhpHtmlDecorator;
     }
 
     /**
@@ -427,7 +427,7 @@ CODE_SAMPLE;
             /** @var Node $previousNode */
             $previousNode = $node->getAttribute(AttributeKey::PREVIOUS_NODE);
 
-            $this->mixPhpHtmlPrinter->before($node);
+            $this->mixPhpHtmlDecorator->decorateBefore($node);
 
             $nodes = [$previousNode, ...$nodes];
         }
@@ -439,7 +439,7 @@ CODE_SAMPLE;
             /** @var Node $nextNode */
             $nextNode = $node->getAttribute(AttributeKey::NEXT_NODE);
 
-            $this->mixPhpHtmlPrinter->after($node, $nodes);
+            $this->mixPhpHtmlDecorator->decorateAfter($node, $nodes);
 
             $nodes = [...$nodes, $nextNode];
         }
